+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Fix `buf push --git-metadata` when local tags point to different objects than
   the remote tags.
 - Fix issue where comment ignores were not respected for `PROTOVALIDATE` lint rule violations.
+- Add `buf beta registry label {create,get,list}` to replace `buf beta registry {draft, tag}`
+  commands.
 
 ## [v1.32.2] - 2024-05-28